MM-8499
First step to getting Jquery kolorpicker working with jquery 1.9+ is to start using the "on"
call.

Previously I was using document level live calls, so the translation is relatively simple.

In the future this code should be rewritten entirely to use local binding rather than top level
events like this.